Fix move of sqltestutil
