--- instrumentation/database/sql/splunksql/internal/sqltestutil/go.mod
@@ -1,30 +1,0 @@
-module github.com/signalfx/splunk-otel-go/instrumentation/database/sql/splunksql/internal/sqltestutil
-
-go 1.18
-
-require (
-	github.com/signalfx/splunk-otel-go v1.4.0
-	github.com/signalfx/splunk-otel-go/instrumentation/database/sql/splunksql v1.4.0
-	github.com/stretchr/testify v1.8.2
-	go.opentelemetry.io/otel v1.14.0
-	go.opentelemetry.io/otel/sdk v1.14.0
-	go.opentelemetry.io/otel/sdk/metric v0.37.0
-)
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/go-logr/logr v1.2.3 // indirect
-	github.com/go-logr/stdr v1.2.2 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/signalfx/splunk-otel-go/instrumentation/internal v1.4.0 // indirect
-	go.opentelemetry.io/otel/metric v0.37.0 // indirect
-	go.opentelemetry.io/otel/trace v1.14.0 // indirect
-	golang.org/x/sys v0.5.0 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-)
-
-replace (
-	github.com/signalfx/splunk-otel-go => ../../../../../../
-	github.com/signalfx/splunk-otel-go/instrumentation/database/sql/splunksql => ../../
-	github.com/signalfx/splunk-otel-go/instrumentation/internal => ../../../../../internal/
-)
