+++ instrumentation/database/sql/splunksql/test/db_metrics_test.go
@@ -21,7 +21,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/signalfx/splunk-otel-go/instrumentation/database/sql/splunksql"
-	"github.com/signalfx/splunk-otel-go/instrumentation/database/sql/splunksql/internal/sqltestutil"
+	"github.com/signalfx/splunk-otel-go/instrumentation/internal/sqltestutil"
 )
 
 func TestMetrics(t *testing.T) {
